Move helper classes to ui module
